pyproject.toml
[project]
name = "news-track"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosmtplib>=4.0.1",
    "apscheduler>=3.11.0",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.178.0",
    "googlesearch-python>=1.3.0",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
]
